move GLU and build_activation in action_type_head.py to ding/torch_utils/network/activation.py, rename them to GLU2 and build_activation2
