+++ ding/torch_utils/network/activation.py
@@ -1,6 +1,7 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+from ding.torch_utils.network.nn_module import fc_block, conv2d_block
 
 
 class GLU(nn.Module):
@@ -88,6 +89,33 @@ def build_activation(activation: str, inplace: bool = None) -> nn.Module:
         # TODO(nyz): influence dizoo/gfootball/model/iql/iql_network.py, and tests of this function
         inplace = True
     act_func = {'relu': nn.ReLU(inplace=inplace), 'glu': GLU, 'prelu': nn.PReLU(), 'swish': Swish()}
+    if activation in act_func.keys():
+        return act_func[activation]
+    else:
+        raise KeyError("invalid key for activation: {}".format(activation))
+
+
+class GLU2(nn.Module):
+    def __init__(self, input_dim, output_dim, context_dim, input_type='fc'):
+        super(GLU2, self).__init__()
+        assert (input_type in ['fc', 'conv2d'])
+        if input_type == 'fc':
+            self.layer1 = fc_block(context_dim, input_dim)
+            self.layer2 = fc_block(input_dim, output_dim)
+        elif input_type == 'conv2d':
+            self.layer1 = conv2d_block(context_dim, input_dim, 1, 1, 0)
+            self.layer2 = conv2d_block(input_dim, output_dim, 1, 1, 0)
+
+    def forward(self, x, context):
+        gate = self.layer1(context)
+        gate = torch.sigmoid(gate)
+        x = gate * x
+        x = self.layer2(x)
+        return x
+
+
+def build_activation2(activation):
+    act_func = {'relu': nn.ReLU(inplace=True), 'glu': GLU2, 'prelu': nn.PReLU(init=0.0)}
     if activation in act_func.keys():
         return act_func[activation]
     else:
